feat: image upload for replies
